+++ Django/robotechnics/partners/pagination.py
@@ -13,7 +13,7 @@ class PartnersPagination(pagination.PageNumberPagination):
     def get_paginated_response(self, data):
         return Response({
             'next': self.get_next_link(),
-            'count': self.page.paginator.count,
             'len': len(data),
+            'total_count': self.page.paginator.count,
             'results': data
         })
