+++ Django/robotechnics/partners/serializers.py
@@ -9,4 +9,4 @@ class PartnerSerializer(serializers.ModelSerializer):
     """
     class Meta:
         model = Partner
-        fields = ['name', 'link_to_the_site', 'photo']
+        fields = ('title', 'link_to_the_site', 'photo')
