+++ Django/robotechnics/robotechnics/urls/api.py
@@ -1,9 +1,9 @@
 from django.urls import include, path
 
 urlpatterns = [
-    path('news', include('news.urls')),
-    path('achievements', include('achievement.urls')),
-    path('partners', include('partners.urls')),
-    path('supervisors', include('director.urls')),
-    path('classic_events/', include('events.urls'))
+    path('news/', include('news.urls')),
+    path('achievements/', include('achievement.urls')),
+    path('partners/', include('partners.urls')),
+    path('supervisors/', include('director.urls')),
+    path('classic_events/', include('events.urls')),
 ]
